🚚 rename: utils 함수 파일이동
