+++ src/shared/utils/truncateString.ts
